wip: refactor logging to config workspace
