+++ config/Cargo.toml
@@ -12,3 +12,20 @@ reqwest = { workspace = true }
 color-eyre = { workspace = true }
 serde = { workspace = true }
 toml = "^0"
+libc = "^0"
+log = "^0"
+tracing = "^0"
+tracing-error = "^0"
+lazy_static = "^1"
+directories = "^5"
+tracing-subscriber = { version = "^0", features = ["env-filter", "serde"] }
+clap = { version = "^4", features = [
+  "derive",
+  "cargo",
+  "wrap_help",
+  "unicode",
+  "string",
+  "unstable-styles",
+] }
+better-panic = "^0"
+strip-ansi-escapes = "^0"
