+++ shared/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod ro_cell;
 pub mod tui;
-pub mod utils;
-pub use log;
 pub use ro_cell::RoCell;
